requirements.txt
charset-normalizer==2.0.10
Django==3.2.11
django-cors-headers==3.11.0
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
phonenumbers==8.12.41
pipreqs==0.4.11
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8